Done with JS test updates and fixtures and snapshots
