+++ app/assets/javascripts/student_profile/LightProfilePage.story.js
@@ -8,7 +8,7 @@ import {
   testPropsForPlutoPoppins,
   testPropsForOlafWhite,
   testPropsForAladdinMouse
-} from './StudentProfilePage.test';
+} from './profileTestProps.fixture';
 
 
 function storifyProps(props) {
