+++ app/assets/javascripts/student_profile/fixtures/serializedDataForAladdinMouse.fixture.js
@@ -1,1 +1,1 @@
-export default {"attendanceData":{"discipline_incidents":[{"id":255,"incident_code":"VIOL-Threat Intimid","created_at":"2018-06-14T17:32:28.359Z","updated_at":"2018-06-14T17:32:28.359Z","incident_location":"Unknown","incident_description":"Earlier today something happened with the student that was concerning and we did this to de-escalate and support them in the moment, and then took these steps to understand the underlying challenges they're facing, and adapt how we're working with them moving forward.","occurred_at":"2017-08-30T13:33:10.711Z","has_exact_time":true,"student_id":104},{"id":254,"incident_code":"Alcohol-Possession","created_at":"2018-06-14T17:32:28.355Z","updated_at":"2018-06-14T17:32:28.355Z","incident_location":"Off campus","incident_description":"Earlier today something happened with the student that was concerning and we did this to de-escalate and support them in the moment, and then took these steps to understand the underlying challenges they're facing, and adapt how we're working with them moving forward.","occurred_at":"2016-07-14T10:13:23.861Z","has_exact_time":true,"student_id":104}],"tardies":[{"id":695,"occurred_at":"2017-09-14T12:56:59.138Z","created_at":"2018-06-14T17:32:28.350Z","updated_at":"2018-06-14T17:32:28.350Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":687,"occurred_at":"2017-08-24T02:21:02.006Z","created_at":"2018-06-14T17:32:28.322Z","updated_at":"2018-06-14T17:32:28.322Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":691,"occurred_at":"2017-01-16T12:11:14.075Z","created_at":"2018-06-14T17:32:28.337Z","updated_at":"2018-06-14T17:32:28.337Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":690,"occurred_at":"2016-07-10T20:17:15.002Z","created_at":"2018-06-14T17:32:28.335Z","updated_at":"2018-06-14T17:32:28.335Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":686,"occurred_at":"2015-11-13T17:06:23.964Z","created_at":"2018-06-14T17:32:28.319Z","updated_at":"2018-06-14T17:32:28.319Z","student_id":104,"dismissed":false,"excused":false,"reason":"Nurse","comment":null},{"id":688,"occurred_at":"2015-08-10T10:09:02.115Z","created_at":"2018-06-14T17:32:28.327Z","updated_at":"2018-06-14T17:32:28.327Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":685,"occurred_at":"2015-02-25T10:06:27.337Z","created_at":"2018-06-14T17:32:28.302Z","updated_at":"2018-06-14T17:32:28.302Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":689,"occurred_at":"2015-01-14T13:09:53.587Z","created_at":"2018-06-14T17:32:28.332Z","updated_at":"2018-06-14T17:32:28.332Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":692,"occurred_at":"2014-10-08T03:39:45.431Z","created_at":"2018-06-14T17:32:28.340Z","updated_at":"2018-06-14T17:32:28.340Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":693,"occurred_at":"2014-09-03T12:18:14.928Z","created_at":"2018-06-14T17:32:28.344Z","updated_at":"2018-06-14T17:32:28.344Z","student_id":104,"dismissed":false,"excused":false,"reason":"Walked","comment":null},{"id":694,"occurred_at":"2013-11-20T09:55:35.281Z","created_at":"2018-06-14T17:32:28.347Z","updated_at":"2018-06-14T17:32:28.347Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null}],"absences":[{"id":695,"occurred_at":"2018-03-23T15:01:34.183Z","created_at":"2018-06-14T17:32:28.298Z","updated_at":"2018-06-14T17:32:28.298Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":694,"occurred_at":"2018-02-06T16:24:18.593Z","created_at":"2018-06-14T17:32:28.294Z","updated_at":"2018-06-14T17:32:28.294Z","student_id":104,"dismissed":false,"excused":false,"reason":"Appointment","comment":null},{"id":703,"occurred_at":"2018-01-24T05:56:26.078Z","created_at":"2018-06-14T17:32:28.352Z","updated_at":"2018-06-14T17:32:28.352Z","student_id":104,"dismissed":false,"excused":false,"reason":"Appointment","comment":null},{"id":700,"occurred_at":"2017-09-30T23:44:29.796Z","created_at":"2018-06-14T17:32:28.324Z","updated_at":"2018-06-14T17:32:28.324Z","student_id":104,"dismissed":false,"excused":false,"reason":"Truant","comment":"Mom called, did not leave a reason."},{"id":701,"occurred_at":"2017-05-02T04:50:49.682Z","created_at":"2018-06-14T17:32:28.330Z","updated_at":"2018-06-14T17:32:28.330Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":696,"occurred_at":"2015-10-19T23:52:28.162Z","created_at":"2018-06-14T17:32:28.306Z","updated_at":"2018-06-14T17:32:28.306Z","student_id":104,"dismissed":false,"excused":false,"reason":"Walked","comment":null},{"id":697,"occurred_at":"2015-09-07T23:32:20.238Z","created_at":"2018-06-14T17:32:28.308Z","updated_at":"2018-06-14T17:32:28.308Z","student_id":104,"dismissed":false,"excused":false,"reason":"Truant","comment":null},{"id":699,"occurred_at":"2015-05-05T21:22:47.196Z","created_at":"2018-06-14T17:32:28.314Z","updated_at":"2018-06-14T17:32:28.314Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":693,"occurred_at":"2015-03-04T21:00:43.509Z","created_at":"2018-06-14T17:32:28.291Z","updated_at":"2018-06-14T17:32:28.291Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":"Mom called, did not leave a reason."},{"id":698,"occurred_at":"2014-05-30T01:04:57.910Z","created_at":"2018-06-14T17:32:28.312Z","updated_at":"2018-06-14T17:32:28.312Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":702,"occurred_at":"2013-07-21T22:42:14.450Z","created_at":"2018-06-14T17:32:28.342Z","updated_at":"2018-06-14T17:32:28.342Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null}]},"currentEducatorAllowedSections":[1,2,3,4,5,6],"sections":[{"id":5,"section_number":"SCI-201A","term_local_id":"S1","schedule":"3(M,W,F)","room_number":"306W","created_at":"2018-06-14T17:31:47.849Z","updated_at":"2018-06-14T17:31:47.849Z","course_id":3,"grade_numeric":"71.0","course_description":"PHYSICS 1","educators":[{"full_name":"Teacher, Fatima"}]}],"transitionNotes":[],"profileInsights": [],"access":{"composite":"6","comprehension":null,"literacy":null,"oral":null,"listening":null,"reading":null,"speaking":null,"writing":null},"educatorsIndex":{"1":{"id":1,"email":"rich@demo.studentinsights.org","full_name":"Districtwide, Rich"},"2":{"id":2,"email":"vivian@demo.studentinsights.org","full_name":"Teacher, Vivian"},"3":{"id":3,"email":"alonso@demo.studentinsights.org","full_name":"Teacher, Alonso"},"4":{"id":4,"email":"silva@demo.studentinsights.org","full_name":"Teacher, Silva"},"5":{"id":5,"email":"laura@demo.studentinsights.org","full_name":"Principal, Laura"},"6":{"id":6,"email":"sarah@demo.studentinsights.org","full_name":"Teacher, Sarah"},"7":{"id":7,"email":"marcus@demo.studentinsights.org","full_name":"Teacher, Marcus"},"8":{"id":8,"email":"les@demo.studentinsights.org","full_name":"Counselor, Les"},"9":{"id":9,"email":"sofia@demo.studentinsights.org","full_name":"Counselor, Sofia"},"10":{"id":10,"email":"jodi@demo.studentinsights.org","full_name":"Teacher, Jodi"},"11":{"id":11,"email":"harry@demo.studentinsights.org","full_name":"Housemaster, Harry"},"12":{"id":12,"email":"bill@demo.studentinsights.org","full_name":"Teacher, Bill"},"13":{"id":13,"email":"hugo@demo.studentinsights.org","full_name":"Teacher, Hugo"},"14":{"id":14,"email":"fatima@demo.studentinsights.org","full_name":"Teacher, Fatima"},"999999":{"id":999999,"email":"uri@demo.studentinsights.org","full_name":"Disney, Uri"}},"serviceTypesIndex":{"502":{"id":502,"name":"Attendance Officer"},"503":{"id":503,"name":"Attendance Contract"},"504":{"id":504,"name":"Behavior Contract"},"505":{"id":505,"name":"Counseling, in-house"},"506":{"id":506,"name":"Counseling, outside"},"507":{"id":507,"name":"Reading intervention"},"508":{"id":508,"name":"Math intervention"},"509":{"id":509,"name":"SomerSession"},"510":{"id":510,"name":"Summer Program for English Language Learners"},"511":{"id":511,"name":"Afterschool Tutoring"},"512":{"id":512,"name":"Freedom School"},"513":{"id":513,"name":"Community Schools"},"514":{"id":514,"name":"X-Block"}},"dibels":[],"chartData":{"star_series_math_percentile":null,"star_series_reading_percentile":null,"next_gen_mcas_mathematics_scaled":null,"next_gen_mcas_ela_scaled":null,"mcas_series_math_scaled":null,"mcas_series_ela_scaled":null,"mcas_series_math_growth":null,"mcas_series_ela_growth":null,"interventions":null},"feed":{"event_notes":[{"id":455,"student_id":104,"educator_id":1,"event_note_type_id":301,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2010-10-20T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":456,"student_id":104,"educator_id":1,"event_note_type_id":305,"text":"Aladdin will go through the special ed process for placement","recorded_at":"2010-11-28T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":457,"student_id":104,"educator_id":1,"event_note_type_id":304,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-01-20T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":458,"student_id":104,"educator_id":1,"event_note_type_id":305,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-02-23T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":459,"student_id":104,"educator_id":1,"event_note_type_id":306,"text":"Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.","recorded_at":"2011-03-29T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":460,"student_id":104,"educator_id":1,"event_note_type_id":301,"text":"Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.","recorded_at":"2011-05-23T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":461,"student_id":104,"educator_id":1,"event_note_type_id":306,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-07-18T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]}],"services":{"active":[],"discontinued":[]},"deprecated":{"interventions":[]}},"student":{"id":104,"has_photo":true,"grade":"10","hispanic_latino":false,"race":"Black","free_reduced_lunch":null,"created_at":"2018-06-14T17:32:28.287Z","updated_at":"2018-06-14T17:32:28.404Z","homeroom_id":4,"first_name":"Aladdin","last_name":"Mouse","state_id":null,"home_language":"Haitian-Creole","school_id":9,"student_address":null,"registration_date":null,"local_id":"000857","program_assigned":"Sp Ed","sped_placement":"Private Separate","disability":"Emotional","sped_level_of_need":"Low >= 2","plan_504":"504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":null,"most_recent_mcas_ela_growth":null,"most_recent_mcas_math_performance":null,"most_recent_mcas_ela_performance":null,"most_recent_mcas_math_scaled":null,"most_recent_mcas_ela_scaled":null,"most_recent_star_reading_percentile":null,"most_recent_star_math_percentile":null,"enrollment_status":"Active","date_of_birth":"2004-05-14T00:00:00.000Z","gender":"M","primary_phone":"999-999-9999 C-Mom","primary_email":"student@example.com","house":"Beacon","counselor":"BUZZ","sped_liaison":null,"absences_count":4,"tardies_count":2,"school_name":"Somerville High","school_local_id":"SHS","school_type":"HS","homeroom_name":"SHS ALL","discipline_incidents_count":1,"restricted_notes_count":0},"currentEducator":{"id":9,"email":"sofia@demo.studentinsights.org","created_at":"2018-06-14T17:31:47.669Z","updated_at":"2018-06-14T23:57:40.360Z","admin":false,"phone":null,"full_name":"Counselor, Sofia","state_id":null,"local_id":null,"staff_type":null,"school_id":9,"schoolwide_access":true,"grade_level_access":[],"restricted_to_sped_students":false,"restricted_to_english_language_learners":false,"can_view_restricted_notes":false,"districtwide_access":false,"can_set_districtwide_access":false,"labels":["use_counselor_based_feed"]}};
+export default {"attendanceData":{"discipline_incidents":[{"id":255,"incident_code":"VIOL-Threat Intimid","created_at":"2018-06-14T17:32:28.359Z","updated_at":"2018-06-14T17:32:28.359Z","incident_location":"Unknown","incident_description":"Earlier today something happened with the student that was concerning and we did this to de-escalate and support them in the moment, and then took these steps to understand the underlying challenges they're facing, and adapt how we're working with them moving forward.","occurred_at":"2017-08-30T13:33:10.711Z","has_exact_time":true,"student_id":104},{"id":254,"incident_code":"Alcohol-Possession","created_at":"2018-06-14T17:32:28.355Z","updated_at":"2018-06-14T17:32:28.355Z","incident_location":"Off campus","incident_description":"Earlier today something happened with the student that was concerning and we did this to de-escalate and support them in the moment, and then took these steps to understand the underlying challenges they're facing, and adapt how we're working with them moving forward.","occurred_at":"2016-07-14T10:13:23.861Z","has_exact_time":true,"student_id":104}],"tardies":[{"id":695,"occurred_at":"2017-09-14T12:56:59.138Z","created_at":"2018-06-14T17:32:28.350Z","updated_at":"2018-06-14T17:32:28.350Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":687,"occurred_at":"2017-08-24T02:21:02.006Z","created_at":"2018-06-14T17:32:28.322Z","updated_at":"2018-06-14T17:32:28.322Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":691,"occurred_at":"2017-01-16T12:11:14.075Z","created_at":"2018-06-14T17:32:28.337Z","updated_at":"2018-06-14T17:32:28.337Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":690,"occurred_at":"2016-07-10T20:17:15.002Z","created_at":"2018-06-14T17:32:28.335Z","updated_at":"2018-06-14T17:32:28.335Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":686,"occurred_at":"2015-11-13T17:06:23.964Z","created_at":"2018-06-14T17:32:28.319Z","updated_at":"2018-06-14T17:32:28.319Z","student_id":104,"dismissed":false,"excused":false,"reason":"Nurse","comment":null},{"id":688,"occurred_at":"2015-08-10T10:09:02.115Z","created_at":"2018-06-14T17:32:28.327Z","updated_at":"2018-06-14T17:32:28.327Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":685,"occurred_at":"2015-02-25T10:06:27.337Z","created_at":"2018-06-14T17:32:28.302Z","updated_at":"2018-06-14T17:32:28.302Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":689,"occurred_at":"2015-01-14T13:09:53.587Z","created_at":"2018-06-14T17:32:28.332Z","updated_at":"2018-06-14T17:32:28.332Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":692,"occurred_at":"2014-10-08T03:39:45.431Z","created_at":"2018-06-14T17:32:28.340Z","updated_at":"2018-06-14T17:32:28.340Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":693,"occurred_at":"2014-09-03T12:18:14.928Z","created_at":"2018-06-14T17:32:28.344Z","updated_at":"2018-06-14T17:32:28.344Z","student_id":104,"dismissed":false,"excused":false,"reason":"Walked","comment":null},{"id":694,"occurred_at":"2013-11-20T09:55:35.281Z","created_at":"2018-06-14T17:32:28.347Z","updated_at":"2018-06-14T17:32:28.347Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null}],"absences":[{"id":695,"occurred_at":"2018-03-23T15:01:34.183Z","created_at":"2018-06-14T17:32:28.298Z","updated_at":"2018-06-14T17:32:28.298Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":694,"occurred_at":"2018-02-06T16:24:18.593Z","created_at":"2018-06-14T17:32:28.294Z","updated_at":"2018-06-14T17:32:28.294Z","student_id":104,"dismissed":false,"excused":false,"reason":"Appointment","comment":null},{"id":703,"occurred_at":"2018-01-24T05:56:26.078Z","created_at":"2018-06-14T17:32:28.352Z","updated_at":"2018-06-14T17:32:28.352Z","student_id":104,"dismissed":false,"excused":false,"reason":"Appointment","comment":null},{"id":700,"occurred_at":"2017-09-30T23:44:29.796Z","created_at":"2018-06-14T17:32:28.324Z","updated_at":"2018-06-14T17:32:28.324Z","student_id":104,"dismissed":false,"excused":false,"reason":"Truant","comment":"Mom called, did not leave a reason."},{"id":701,"occurred_at":"2017-05-02T04:50:49.682Z","created_at":"2018-06-14T17:32:28.330Z","updated_at":"2018-06-14T17:32:28.330Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":696,"occurred_at":"2015-10-19T23:52:28.162Z","created_at":"2018-06-14T17:32:28.306Z","updated_at":"2018-06-14T17:32:28.306Z","student_id":104,"dismissed":false,"excused":false,"reason":"Walked","comment":null},{"id":697,"occurred_at":"2015-09-07T23:32:20.238Z","created_at":"2018-06-14T17:32:28.308Z","updated_at":"2018-06-14T17:32:28.308Z","student_id":104,"dismissed":false,"excused":false,"reason":"Truant","comment":null},{"id":699,"occurred_at":"2015-05-05T21:22:47.196Z","created_at":"2018-06-14T17:32:28.314Z","updated_at":"2018-06-14T17:32:28.314Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":693,"occurred_at":"2015-03-04T21:00:43.509Z","created_at":"2018-06-14T17:32:28.291Z","updated_at":"2018-06-14T17:32:28.291Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":"Mom called, did not leave a reason."},{"id":698,"occurred_at":"2014-05-30T01:04:57.910Z","created_at":"2018-06-14T17:32:28.312Z","updated_at":"2018-06-14T17:32:28.312Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null},{"id":702,"occurred_at":"2013-07-21T22:42:14.450Z","created_at":"2018-06-14T17:32:28.342Z","updated_at":"2018-06-14T17:32:28.342Z","student_id":104,"dismissed":false,"excused":false,"reason":null,"comment":null}]},"currentEducatorAllowedSections":[1,2,3,4,5,6],"sections":[{"id":5,"section_number":"SCI-201A","term_local_id":"S1","schedule":"3(M,W,F)","room_number":"306W","created_at":"2018-06-14T17:31:47.849Z","updated_at":"2018-06-14T17:31:47.849Z","course_id":3,"grade_numeric":"71.0","course_description":"PHYSICS 1","educators":[{"full_name":"Teacher, Fatima"}]}],"transitionNotes":[],"profileInsights": [],"access":{"composite":null,"comprehension":null,"literacy":null,"oral":null,"listening":null,"reading":null,"speaking":null,"writing":null},"educatorsIndex":{"1":{"id":1,"email":"rich@demo.studentinsights.org","full_name":"Districtwide, Rich"},"2":{"id":2,"email":"vivian@demo.studentinsights.org","full_name":"Teacher, Vivian"},"3":{"id":3,"email":"alonso@demo.studentinsights.org","full_name":"Teacher, Alonso"},"4":{"id":4,"email":"silva@demo.studentinsights.org","full_name":"Teacher, Silva"},"5":{"id":5,"email":"laura@demo.studentinsights.org","full_name":"Principal, Laura"},"6":{"id":6,"email":"sarah@demo.studentinsights.org","full_name":"Teacher, Sarah"},"7":{"id":7,"email":"marcus@demo.studentinsights.org","full_name":"Teacher, Marcus"},"8":{"id":8,"email":"les@demo.studentinsights.org","full_name":"Counselor, Les"},"9":{"id":9,"email":"sofia@demo.studentinsights.org","full_name":"Counselor, Sofia"},"10":{"id":10,"email":"jodi@demo.studentinsights.org","full_name":"Teacher, Jodi"},"11":{"id":11,"email":"harry@demo.studentinsights.org","full_name":"Housemaster, Harry"},"12":{"id":12,"email":"bill@demo.studentinsights.org","full_name":"Teacher, Bill"},"13":{"id":13,"email":"hugo@demo.studentinsights.org","full_name":"Teacher, Hugo"},"14":{"id":14,"email":"fatima@demo.studentinsights.org","full_name":"Teacher, Fatima"},"999999":{"id":999999,"email":"uri@demo.studentinsights.org","full_name":"Disney, Uri"}},"serviceTypesIndex":{"502":{"id":502,"name":"Attendance Officer"},"503":{"id":503,"name":"Attendance Contract"},"504":{"id":504,"name":"Behavior Contract"},"505":{"id":505,"name":"Counseling, in-house"},"506":{"id":506,"name":"Counseling, outside"},"507":{"id":507,"name":"Reading intervention"},"508":{"id":508,"name":"Math intervention"},"509":{"id":509,"name":"SomerSession"},"510":{"id":510,"name":"Summer Program for English Language Learners"},"511":{"id":511,"name":"Afterschool Tutoring"},"512":{"id":512,"name":"Freedom School"},"513":{"id":513,"name":"Community Schools"},"514":{"id":514,"name":"X-Block"}},"dibels":[],"chartData":{"star_series_math_percentile":null,"star_series_reading_percentile":null,"next_gen_mcas_mathematics_scaled":null,"next_gen_mcas_ela_scaled":null,"mcas_series_math_scaled":null,"mcas_series_ela_scaled":null,"mcas_series_math_growth":null,"mcas_series_ela_growth":null,"interventions":null},"feed":{"event_notes":[{"id":455,"student_id":104,"educator_id":1,"event_note_type_id":301,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2010-10-20T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":456,"student_id":104,"educator_id":1,"event_note_type_id":305,"text":"Aladdin will go through the special ed process for placement","recorded_at":"2010-11-28T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":457,"student_id":104,"educator_id":1,"event_note_type_id":304,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-01-20T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":458,"student_id":104,"educator_id":1,"event_note_type_id":305,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-02-23T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":459,"student_id":104,"educator_id":1,"event_note_type_id":306,"text":"Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.","recorded_at":"2011-03-29T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":460,"student_id":104,"educator_id":1,"event_note_type_id":301,"text":"Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.","recorded_at":"2011-05-23T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":461,"student_id":104,"educator_id":1,"event_note_type_id":306,"text":"Philis is meeting with parent next week and will present an attendance contract.","recorded_at":"2011-07-18T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]}],"services":{"active":[],"discontinued":[]},"deprecated":{"interventions":[]}},"student":{"id":104,"has_photo":true,"grade":"10","hispanic_latino":false,"race":"Black","free_reduced_lunch":null,"created_at":"2018-06-14T17:32:28.287Z","updated_at":"2018-06-14T17:32:28.404Z","homeroom_id":4,"first_name":"Aladdin","last_name":"Mouse","state_id":null,"home_language":"Haitian-Creole","school_id":9,"student_address":null,"registration_date":null,"local_id":"000857","program_assigned":"Sp Ed","sped_placement":"Private Separate","disability":"Emotional","sped_level_of_need":"Low >= 2","plan_504":"504","limited_english_proficiency":"FLEP","most_recent_mcas_math_growth":null,"most_recent_mcas_ela_growth":null,"most_recent_mcas_math_performance":null,"most_recent_mcas_ela_performance":null,"most_recent_mcas_math_scaled":null,"most_recent_mcas_ela_scaled":null,"most_recent_star_reading_percentile":null,"most_recent_star_math_percentile":null,"enrollment_status":"Active","date_of_birth":"2004-05-14T00:00:00.000Z","gender":"M","primary_phone":"999-999-9999 C-Mom","primary_email":"student@example.com","house":"Beacon","counselor":"BUZZ","sped_liaison":null,"absences_count":4,"tardies_count":2,"school_name":"Somerville High","school_local_id":"SHS","school_type":"HS","homeroom_name":"SHS ALL","discipline_incidents_count":1,"restricted_notes_count":0},"currentEducator":{"id":9,"email":"sofia@demo.studentinsights.org","created_at":"2018-06-14T17:31:47.669Z","updated_at":"2018-06-14T23:57:40.360Z","admin":false,"phone":null,"full_name":"Counselor, Sofia","state_id":null,"local_id":null,"staff_type":null,"school_id":9,"schoolwide_access":true,"grade_level_access":[],"restricted_to_sped_students":false,"restricted_to_english_language_learners":false,"can_view_restricted_notes":false,"districtwide_access":false,"can_set_districtwide_access":false,"labels":["use_counselor_based_feed"]}};
