+++ app/assets/javascripts/student_profile/fixtures/serializedDataForPlutoPoppins.fixture.js
@@ -1,1 +1,1 @@
-export default {"attendanceData":{"discipline_incidents":[],"tardies":[],"absences":[{"id":47,"occurred_at":"2017-01-16T22:06:51.248Z","created_at":"2018-06-14T14:18:20.343Z","updated_at":"2018-06-14T14:18:20.343Z","student_id":8,"dismissed":false,"excused":false,"reason":"Medical","comment":null}]},"currentEducatorAllowedSections":[1,2,3,4,5,6],"sections":[],"transitionNotes":[],"profileInsights": [],"access":{"composite":"9","comprehension":null,"literacy":null,"oral":null,"listening":null,"reading":null,"speaking":null,"writing":null},"educatorsIndex":{"1":{"id":1,"email":"rich@demo.studentinsights.org","full_name":"Districtwide, Rich"},"2":{"id":2,"email":"vivian@demo.studentinsights.org","full_name":"Teacher, Vivian"},"3":{"id":3,"email":"alonso@demo.studentinsights.org","full_name":"Teacher, Alonso"},"4":{"id":4,"email":"silva@demo.studentinsights.org","full_name":"Teacher, Silva"},"5":{"id":5,"email":"laura@demo.studentinsights.org","full_name":"Principal, Laura"},"6":{"id":6,"email":"sarah@demo.studentinsights.org","full_name":"Teacher, Sarah"},"7":{"id":7,"email":"marcus@demo.studentinsights.org","full_name":"Teacher, Marcus"},"8":{"id":8,"email":"les@demo.studentinsights.org","full_name":"Counselor, Les"},"9":{"id":9,"email":"sofia@demo.studentinsights.org","full_name":"Counselor, Sofia"},"10":{"id":10,"email":"jodi@demo.studentinsights.org","full_name":"Teacher, Jodi"},"11":{"id":11,"email":"harry@demo.studentinsights.org","full_name":"Housemaster, Harry"},"12":{"id":12,"email":"bill@demo.studentinsights.org","full_name":"Teacher, Bill"},"13":{"id":13,"email":"hugo@demo.studentinsights.org","full_name":"Teacher, Hugo"},"14":{"id":14,"email":"fatima@demo.studentinsights.org","full_name":"Teacher, Fatima"},"999999":{"id":999999,"email":"uri@demo.studentinsights.org","full_name":"Disney, Uri"}},"serviceTypesIndex":{"502":{"id":502,"name":"Attendance Officer"},"503":{"id":503,"name":"Attendance Contract"},"504":{"id":504,"name":"Behavior Contract"},"505":{"id":505,"name":"Counseling, in-house"},"506":{"id":506,"name":"Counseling, outside"},"507":{"id":507,"name":"Reading intervention"},"508":{"id":508,"name":"Math intervention"},"509":{"id":509,"name":"SomerSession"},"510":{"id":510,"name":"Summer Program for English Language Learners"},"511":{"id":511,"name":"Afterschool Tutoring"},"512":{"id":512,"name":"Freedom School"},"513":{"id":513,"name":"Community Schools"},"514":{"id":514,"name":"X-Block"}},"dibels":[{"benchmark":"INTENSIVE","date_taken":"2015-05-15T00:00:00.000Z"},{"benchmark":"INTENSIVE","date_taken":"2013-05-15T00:00:00.000Z"},{"benchmark":"STRATEGIC","date_taken":"2014-05-15T00:00:00.000Z"},{"benchmark":"STRATEGIC","date_taken":"2011-05-15T00:00:00.000Z"}],"chartData":{"star_series_math_percentile":[{id: 1, date_taken: '2018-02-13T22:17:30.338Z', percentile_rank:10,grade_equivalent:'1.00',total_time:1400}],"star_series_reading_percentile":[{id: 1,date_taken:'2018-02-13T22:17:30.338Z',percentile_rank: 10, grade_equivalent:'1.00',total_time:1300}],"next_gen_mcas_mathematics_scaled":[[2010,5,15,441],[2011,5,15,449],[2014,5,15,465],[2015,5,15,532]],"next_gen_mcas_ela_scaled":[[2010,5,15,412],[2012,5,15,433],[2013,5,15,558],[2014,5,15,447]],"mcas_series_math_scaled":[[2010,5,15,222],[2012,5,15,278],[2013,5,15,206],[2014,5,15,256]],"mcas_series_ela_scaled":[[2011,5,15,251],[2013,5,15,225],[2014,5,15,248],[2015,5,15,253]],"mcas_series_math_growth":[[2010,5,15,93],[2012,5,15,45],[2013,5,15,38],[2014,5,15,90]],"mcas_series_ela_growth":[[2011,5,15,79],[2013,5,15,50],[2014,5,15,71],[2015,5,15,57]],"interventions":null},"feed":{"event_notes":[{"id":23,"student_id":8,"educator_id":1,"event_note_type_id":302,"text":"<redacted>","recorded_at":"2010-10-30T00:00:00.000Z","is_restricted":true,"event_note_revisions_count":0,"attachments":[]},{"id":24,"student_id":8,"educator_id":1,"event_note_type_id":302,"text":"Lila will look into a referral to a big sister program.","recorded_at":"2010-12-02T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":25,"student_id":8,"educator_id":1,"event_note_type_id":305,"text":"Lila will look into a referral to a big sister program.","recorded_at":"2011-01-31T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":26,"student_id":8,"educator_id":1,"event_note_type_id":300,"text":"Meetings with staff about lack of academic progress.","recorded_at":"2011-03-10T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":27,"student_id":8,"educator_id":1,"event_note_type_id":304,"text":"Meetings with staff about lack of academic progress.","recorded_at":"2011-04-21T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]}],"services":{"active":[{"id":3,"student_id":8,"provided_by_educator_name":"Districtwide, Rich","service_type_id":507,"date_started":"2018-02-20T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":8,"recorded_at":"2018-02-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":4,"student_id":8,"provided_by_educator_name":"Districtwide, Rich","service_type_id":514,"date_started":"2018-03-11T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":13,"recorded_at":"2018-03-11T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":5,"student_id":8,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2018-01-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2018-01-14T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":6,"student_id":8,"provided_by_educator_name":null,"service_type_id":502,"date_started":"2017-11-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":10,"recorded_at":"2017-11-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":7,"student_id":8,"provided_by_educator_name":null,"service_type_id":504,"date_started":"2017-10-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":9,"recorded_at":"2017-10-17T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"discontinued":[]},"deprecated":{"interventions":[]}},"student":{"id":8,"grade":"KF","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Not Eligible","created_at":"2018-06-14T14:18:20.336Z","updated_at":"2018-06-14T14:19:22.941Z","homeroom_id":1,"first_name":"Pluto","last_name":"Poppins","state_id":null,"home_language":"Haitian-Creole","school_id":2,"student_address":null,"registration_date":null,"local_id":"000681","program_assigned":"Sp Ed","sped_placement":"Partial Inclusion","disability":"Communication","sped_level_of_need":"Low >= 2","plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":51,"most_recent_mcas_ela_growth":57,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"W","most_recent_mcas_math_scaled":532,"most_recent_mcas_ela_scaled":253,"most_recent_star_reading_percentile":34,"most_recent_star_math_percentile":31,"enrollment_status":"Active","date_of_birth":"2013-12-16T00:00:00.000Z","gender":"F","primary_phone":"999-999-9999 C-Mom","primary_email":"student@example.com","house":null,"counselor":null,"sped_liaison":"CONCEPCION","absences_count":0,"tardies_count":0,"school_name":"Arthur D Healey","school_type":"ESMS","homeroom_name":"HEA 003","discipline_incidents_count":0,"restricted_notes_count":0},"currentEducator":{"id":999999,"email":"uri@demo.studentinsights.org","created_at":"2018-06-14T14:18:17.476Z","updated_at":"2018-06-14T14:18:17.476Z","admin":true,"phone":null,"full_name":"Disney, Uri","state_id":null,"local_id":null,"staff_type":"Administrator","school_id":2,"schoolwide_access":true,"grade_level_access":[],"restricted_to_sped_students":false,"restricted_to_english_language_learners":false,"can_view_restricted_notes":true,"districtwide_access":true,"can_set_districtwide_access":true,"labels":[]}};
+export default {"attendanceData":{"discipline_incidents":[],"tardies":[],"absences":[{"id":47,"occurred_at":"2017-01-16T22:06:51.248Z","created_at":"2018-06-14T14:18:20.343Z","updated_at":"2018-06-14T14:18:20.343Z","student_id":8,"dismissed":false,"excused":false,"reason":"Medical","comment":null}]},"currentEducatorAllowedSections":[1,2,3,4,5,6],"sections":[],"transitionNotes":[],"profileInsights": [],"access":{"composite":null,"comprehension":null,"literacy":null,"oral":null,"listening":null,"reading":null,"speaking":null,"writing":null},"educatorsIndex":{"1":{"id":1,"email":"rich@demo.studentinsights.org","full_name":"Districtwide, Rich"},"2":{"id":2,"email":"vivian@demo.studentinsights.org","full_name":"Teacher, Vivian"},"3":{"id":3,"email":"alonso@demo.studentinsights.org","full_name":"Teacher, Alonso"},"4":{"id":4,"email":"silva@demo.studentinsights.org","full_name":"Teacher, Silva"},"5":{"id":5,"email":"laura@demo.studentinsights.org","full_name":"Principal, Laura"},"6":{"id":6,"email":"sarah@demo.studentinsights.org","full_name":"Teacher, Sarah"},"7":{"id":7,"email":"marcus@demo.studentinsights.org","full_name":"Teacher, Marcus"},"8":{"id":8,"email":"les@demo.studentinsights.org","full_name":"Counselor, Les"},"9":{"id":9,"email":"sofia@demo.studentinsights.org","full_name":"Counselor, Sofia"},"10":{"id":10,"email":"jodi@demo.studentinsights.org","full_name":"Teacher, Jodi"},"11":{"id":11,"email":"harry@demo.studentinsights.org","full_name":"Housemaster, Harry"},"12":{"id":12,"email":"bill@demo.studentinsights.org","full_name":"Teacher, Bill"},"13":{"id":13,"email":"hugo@demo.studentinsights.org","full_name":"Teacher, Hugo"},"14":{"id":14,"email":"fatima@demo.studentinsights.org","full_name":"Teacher, Fatima"},"999999":{"id":999999,"email":"uri@demo.studentinsights.org","full_name":"Disney, Uri"}},"serviceTypesIndex":{"502":{"id":502,"name":"Attendance Officer"},"503":{"id":503,"name":"Attendance Contract"},"504":{"id":504,"name":"Behavior Contract"},"505":{"id":505,"name":"Counseling, in-house"},"506":{"id":506,"name":"Counseling, outside"},"507":{"id":507,"name":"Reading intervention"},"508":{"id":508,"name":"Math intervention"},"509":{"id":509,"name":"SomerSession"},"510":{"id":510,"name":"Summer Program for English Language Learners"},"511":{"id":511,"name":"Afterschool Tutoring"},"512":{"id":512,"name":"Freedom School"},"513":{"id":513,"name":"Community Schools"},"514":{"id":514,"name":"X-Block"}},"dibels":[{"benchmark":"INTENSIVE","date_taken":"2015-05-15T00:00:00.000Z"},{"benchmark":"INTENSIVE","date_taken":"2013-05-15T00:00:00.000Z"},{"benchmark":"STRATEGIC","date_taken":"2014-05-15T00:00:00.000Z"},{"benchmark":"STRATEGIC","date_taken":"2011-05-15T00:00:00.000Z"}],"chartData":{"star_series_math_percentile":[{id: 1, date_taken: '2018-02-13T22:17:30.338Z', percentile_rank:10,grade_equivalent:'1.00',total_time:1400}],"star_series_reading_percentile":[{id: 1,date_taken:'2018-02-13T22:17:30.338Z',percentile_rank: 10, grade_equivalent:'1.00',total_time:1300}],"next_gen_mcas_mathematics_scaled":[[2010,5,15,441],[2011,5,15,449],[2014,5,15,465],[2015,5,15,532]],"next_gen_mcas_ela_scaled":[[2010,5,15,412],[2012,5,15,433],[2013,5,15,558],[2014,5,15,447]],"mcas_series_math_scaled":[[2010,5,15,222],[2012,5,15,278],[2013,5,15,206],[2014,5,15,256]],"mcas_series_ela_scaled":[[2011,5,15,251],[2013,5,15,225],[2014,5,15,248],[2015,5,15,253]],"mcas_series_math_growth":[[2010,5,15,93],[2012,5,15,45],[2013,5,15,38],[2014,5,15,90]],"mcas_series_ela_growth":[[2011,5,15,79],[2013,5,15,50],[2014,5,15,71],[2015,5,15,57]],"interventions":null},"feed":{"event_notes":[{"id":23,"student_id":8,"educator_id":1,"event_note_type_id":302,"text":"<redacted>","recorded_at":"2010-10-30T00:00:00.000Z","is_restricted":true,"event_note_revisions_count":0,"attachments":[]},{"id":24,"student_id":8,"educator_id":1,"event_note_type_id":302,"text":"Lila will look into a referral to a big sister program.","recorded_at":"2010-12-02T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":25,"student_id":8,"educator_id":1,"event_note_type_id":305,"text":"Lila will look into a referral to a big sister program.","recorded_at":"2011-01-31T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":26,"student_id":8,"educator_id":1,"event_note_type_id":300,"text":"Meetings with staff about lack of academic progress.","recorded_at":"2011-03-10T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]},{"id":27,"student_id":8,"educator_id":1,"event_note_type_id":304,"text":"Meetings with staff about lack of academic progress.","recorded_at":"2011-04-21T00:00:00.000Z","is_restricted":false,"event_note_revisions_count":0,"attachments":[]}],"services":{"active":[{"id":3,"student_id":8,"provided_by_educator_name":"Districtwide, Rich","service_type_id":507,"date_started":"2018-02-20T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":8,"recorded_at":"2018-02-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":4,"student_id":8,"provided_by_educator_name":"Districtwide, Rich","service_type_id":514,"date_started":"2018-03-11T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":13,"recorded_at":"2018-03-11T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":5,"student_id":8,"provided_by_educator_name":null,"service_type_id":510,"date_started":"2018-01-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":6,"recorded_at":"2018-01-14T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":6,"student_id":8,"provided_by_educator_name":null,"service_type_id":502,"date_started":"2017-11-24T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":10,"recorded_at":"2017-11-24T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null},{"id":7,"student_id":8,"provided_by_educator_name":null,"service_type_id":504,"date_started":"2017-10-14T00:00:00.000Z","estimated_end_date":null,"recorded_by_educator_id":9,"recorded_at":"2017-10-17T00:00:00.000Z","discontinued_by_educator_id":null,"discontinued_recorded_at":null}],"discontinued":[]},"deprecated":{"interventions":[]}},"student":{"id":8,"grade":"KF","hispanic_latino":true,"race":"Asian","free_reduced_lunch":"Not Eligible","created_at":"2018-06-14T14:18:20.336Z","updated_at":"2018-06-14T14:19:22.941Z","homeroom_id":1,"first_name":"Pluto","last_name":"Poppins","state_id":null,"home_language":"Haitian-Creole","school_id":2,"student_address":null,"registration_date":null,"local_id":"000681","program_assigned":"Sp Ed","sped_placement":"Partial Inclusion","disability":"Communication","sped_level_of_need":"Low >= 2","plan_504":"Not 504","limited_english_proficiency":"Limited","most_recent_mcas_math_growth":51,"most_recent_mcas_ela_growth":57,"most_recent_mcas_math_performance":"Not Meeting Expectations","most_recent_mcas_ela_performance":"W","most_recent_mcas_math_scaled":532,"most_recent_mcas_ela_scaled":253,"most_recent_star_reading_percentile":34,"most_recent_star_math_percentile":31,"enrollment_status":"Active","date_of_birth":"2013-12-16T00:00:00.000Z","gender":"F","primary_phone":"999-999-9999 C-Mom","primary_email":"student@example.com","house":null,"counselor":null,"sped_liaison":"CONCEPCION","absences_count":0,"tardies_count":0,"school_name":"Arthur D Healey","school_type":"ESMS","homeroom_name":"HEA 003","discipline_incidents_count":0,"restricted_notes_count":0},"currentEducator":{"id":999999,"email":"uri@demo.studentinsights.org","created_at":"2018-06-14T14:18:17.476Z","updated_at":"2018-06-14T14:18:17.476Z","admin":true,"phone":null,"full_name":"Disney, Uri","state_id":null,"local_id":null,"staff_type":"Administrator","school_id":2,"schoolwide_access":true,"grade_level_access":[],"restricted_to_sped_students":false,"restricted_to_english_language_learners":false,"can_view_restricted_notes":true,"districtwide_access":true,"can_set_districtwide_access":true,"labels":[]}};
